+++ packages/components/src/styles/componentTypes.ts
@@ -202,6 +202,7 @@ export interface DndStyles {
     dragging: string
     content: string
     removeButton: string
+    iconContainer: string
   },
   droppable: {
     root: string
@@ -213,6 +214,7 @@ export interface DndStyles {
     dragging: string
     content: string
     removeButton: string
+    iconContainer: string
   }
 }