+++ packages/components/src/tooltip/Tooltip.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react'
 import { appliedStyles, cns } from '@stcland/utils'
 import { TooltipStyles } from 'src/styles'
+import useOutsideClick from '../utils'
 
 //*****************************************************************************
 // Interface
@@ -95,6 +96,11 @@ const Tooltip = ({
     }
   }
 
+  useOutsideClick(tooltipRef, {
+    onClickOutside: hideTooltip,
+    capture: false,
+  })
+
   const getTooltipPosition = () => {
     switch (placement) {
     case 'top':
@@ -142,7 +148,7 @@ const Tooltip = ({
   }
 
   const tooltipStyles : TooltipStyles = {
-    root: 'relative inline-block',
+    root: 'relative',
     tooltipContainer: 'absolute z-50 p-3 rounded-md shadow-md text-sm w-max',
     colorClass: 'bg-zinc-600 text-white',
     arrow: 'absolute w-3.5 h-2.5 border-t-2 border-r-2 border-transparent',