+++ packages/components/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @stcland/components
 
+## 0.0.18
+
+### Patch Changes
+
+- Fixing default style in tooltip.
+
+## 0.0.17
+
+### Patch Changes
+
+- Adding icon prop to DND, and outside click in Tooltip.
+
 ## 0.0.16
 
 ### Patch Changes